extract scorer.ts and give OSS attribution
